Added HUD and partially finished support for menu's

+++ DAE16_Tirez_Arthur/EnemyCrucified.h
@@ -20,6 +20,6 @@ private:
 	static const float MAXHEALTH;
 	static const float SPEED;
 	static const float ATTACKDMG;
-	static const float POINTS;
+	static const int POINTS;
 };
